feat: refactor chat to sveltekit routes
